Adds more tests

+++ test/integration/datatype/decimal.js
@@ -9,7 +9,66 @@ module.exports = function (knex) {
       return Promise.resolve();
     }
     const tableName = 'decimal_test';
-    const testDecimal = 10000000000005.74;
+    const testDecimal = 12345678901234.56;
+    return knex
+      .transaction(function (tr) {
+        return tr.schema.dropTableIfExists(tableName).then(function () {
+          return tr.schema.createTable(tableName, function (table) {
+            table.decimal('largeDecimal', 38, 10);
+          });
+        });
+      })
+      .then(function () {
+        return knex(tableName).insert({
+          largeDecimal: testDecimal,
+        });
+      })
+      .then(function () {
+        return knex(tableName).select('*');
+      })
+      .then(function (result) {
+        expect(result[0].largeDecimal).to.equal(testDecimal);
+      })
+      .catch(function (err) {
+        throw new Error('Test should not throw an error');
+      });
+  });
+  it('#test decimal mssql should allow numbers with precision of 10', function () {
+    if (!/mssql/i.test(knex.client.driverName)) {
+      return Promise.resolve();
+    }
+    const tableName = 'decimal_test';
+    const testDecimal = 0.1234567891;
+    return knex
+      .transaction(function (tr) {
+        return tr.schema.dropTableIfExists(tableName).then(function () {
+          return tr.schema.createTable(tableName, function (table) {
+            table.decimal('largeDecimal', 38, 10);
+          });
+        });
+      })
+      .then(function () {
+        return knex(tableName).insert({
+          largeDecimal: testDecimal,
+        });
+      })
+      .then(function () {
+        return knex(tableName).select('*');
+      })
+      .then(function (result) {
+        expect(result[0].largeDecimal).to.equal(testDecimal);
+      })
+      .catch(function (err) {
+        throw new Error('Test should not throw an error');
+      });
+  });
+  it('#test decimal mssql should allow numbers with precision of 16', function () {
+    // Plain JavaScript cannot handle higher numbers without losing precision.
+    if (!/mssql/i.test(knex.client.driverName)) {
+      return Promise.resolve();
+    }
+    const tableName = 'decimal_test';
+    const testDecimal = 1234567890123456;
     return knex
       .transaction(function (tr) {
         return tr.schema.dropTableIfExists(tableName).then(function () {
